Replace benchmark.cpp with clickbench.cpp for enhanced cache prewarm benchmarking. The new implementation supports additional options and improved functionality, including query repetition and dynamic mode selection.
